requirements-dev.txt
flake8
mamba
expects
doublex
doublex-expects
python-coveralls
PyYAML
